% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen2.R
\name{mark_dev_block}
\alias{mark_dev_block}
\title{Mark dev blocks as dev blocks}
\usage{
mark_dev_block(block, env, base_path)
}
\arguments{
\item{block}{The block, a sort of list of tags}

\item{env}{No idea}

\item{base_path}{Path on which roxygenize is run}
}
\description{
This function "marks" blocks with a dev tag
by adding "An internal function: " at the beginning
of their title.
This is needed because there is no other way
to let the output method of the roclet
know which topics "are" dev, because
the output method of the Rd roclet no longer has
access to the tags.
}
