% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen2.R
\name{rbuild_ignore}
\alias{rbuild_ignore}
\title{My helper function}
\usage{
rbuild_ignore(topic, base_path)
}
\arguments{
\item{topic}{A roxygen2 topic}

\item{base_path}{Path on which roxygenize is run}
}
\value{
No idea, the Rd roclet uses it
}
\description{
Internal function: My helper function
}
